+++ app/components/MapOfOodi/messages.js
@@ -1,0 +1,15 @@
+/*
+ * Submit Messages
+ *
+ * This contains all the text for the HomePage component.
+ */
+import { defineMessages } from 'react-intl';
+
+const scope = 'components.MapOfOodi';
+
+export default defineMessages({
+  youAreHere: {
+    id: `${scope}.youAreHere`,
+    defaultMessage: 'Olet tässä',
+  },
+});