+++ app/components/VacancyLabel/index.js
@@ -2,15 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import styled from 'styled-components';
-import classNames from 'classnames';
 
+import Legend from 'components/Legend';
 import SpaceAvailability from 'constants/SpaceAvailability';
 import VacancyIcon from 'components/VacancyIcon';
-import P from 'components/P';
 import messages from './messages';
 
 const VARIANTS = ['dark', 'light'];
 const VACANCIES = Object.values(SpaceAvailability);
+// This is a special state that categorizes SpaceAvailability.NO_DATA
+// and SpaceAvailability.UNKNOWN into a more human friendly concept.
+const VACANCY_NOT_RESERVABLE = 'notReservable';
 
 export const Wrapper = styled.div`
   display: flex;
@@ -23,16 +25,15 @@ export const Wrapper = styled.div`
 `;
 
 const VacancyLabel = ({ vacancy, variant = VARIANTS[0] }) => (
-  <Wrapper>
-    <VacancyIcon className={vacancy} />
-    <P className={classNames(['small', variant])}>
-      <FormattedMessage {...messages[vacancy || 'available']} />
-    </P>
-  </Wrapper>
+  <Legend
+    icon={<VacancyIcon className={vacancy} />}
+    label={<FormattedMessage {...messages[vacancy || 'available']} />}
+    labelClass={variant}
+  />
 );
 
 VacancyLabel.propTypes = {
-  vacancy: PropTypes.oneOf(VACANCIES),
+  vacancy: PropTypes.oneOf([...VACANCIES, VACANCY_NOT_RESERVABLE]),
   variant: PropTypes.oneOf(VARIANTS),
 };