+++ CHANGELOG.md
@@ -5,6 +5,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-06-10
+### Added
+- Add closed state for tooltip groups
+- Add "you are here" label into map legend
+- Add not bookable state into map legend
+- Add special label for GROUP_ROOM_1
+
+### Changed
+- Use pattern instead of solid colour for taken rooms
+- Use a font sizes larger or as large as 15px
+- Change not reservable colour into blue and selected colour into white
+- Don't show availability status for machine room
+- Recording studio button opens tooltip on top of the rightmost room in the recording studio setup
+
+### Fixed
+- Fix untranslated "you are here" label
+- Fix wrong font family for "you are here" label
+- Fix room data for retro Gaming corner, Gaming hall and Kuutio
+- Fix translations coverstitch machine, printing room and language menu
+- Fix naming of Studio 1
+- Fix game room order
+- Fix you are here label overlapping position indicator in English translation
+
 ## [0.1.0] - 2020-04-16
 ### Added
 - Add QR link to Varaamo